New objects helpers
"Invokable" objects are defined as an Abstract base class and a
read-to-use registry class to help working with object allowing
"setProp" and "getProp" magic methods and some more features.

A new Helper has been set to make properties and methods names
transofmation (passing from CamelCase to underscored and back) in
"CodeHelper".

Upgrade to version 1.0.3